fixing levels test case
